added linux startup guide ENG
